+++ apps/hubble/src/cli.ts
@@ -8,7 +8,6 @@ import fs, { existsSync } from "fs";
 import { mkdir, readFile, writeFile } from "fs/promises";
 import { Result, ResultAsync } from "neverthrow";
 import { dirname, resolve } from "path";
-import { exit } from "process";
 import {
   APP_VERSION,
   FARCASTER_VERSION,
@@ -60,6 +59,12 @@ const parseNumber = (string: string) => {
   const number = Number(string);
   if (isNaN(number)) throw new Error("Not a number.");
   return number;
+};
+
+// Use [return flushAndExit(code)] instead of [process.exit(code)] to ensure that logs get written. Prefixing with [return] maintains the type inferencing you get from using [process.exit].
+const flushAndExit = (exitCode: number) => {
+  logger.flush();
+  process.exit(exitCode);
 };
 
 const app = new Command();
@@ -211,16 +216,16 @@ app
           .teardown(shutdownReason)
           .then(() => {
             logger.info("Hub stopped gracefully");
-            process.exit(0);
+            return flushAndExit(0);
           })
           .catch((err) => {
             logger.error({ reason: `Error stopping hub: ${err}` });
-            process.exit(1);
+            return flushAndExit(1);
           });
 
         setTimeout(() => {
           logger.fatal("Forcing exit after grace period");
-          process.exit(1);
+          return flushAndExit(1);
         }, SHUTDOWN_GRACE_PERIOD_MS);
       }
     };
@@ -242,7 +247,7 @@ app
         StartupCheckStatus.ERROR,
         `Hubble requires at least 16GB of RAM to run. Detected ${totalMemory}GB`,
       );
-      process.exit(1);
+      return flushAndExit(1);
     } else {
       startupCheck.printStartupCheckStatus(StartupCheckStatus.OK, `Detected ${totalMemory}GB of RAM`);
     }
@@ -330,7 +335,7 @@ app
           `Failed to read identity from ${cliOptions.id}. Please run "yarn identity create".`,
           "https://www.thehubble.xyz/intro/install.html#installing-hubble\n",
         );
-        process.exit(1);
+        return flushAndExit(1);
       } else {
         peerId = peerIdR.value;
       }
@@ -359,7 +364,7 @@ app
           "https://www.thehubble.xyz/intro/install.html#installing-hubble\n",
         );
 
-        process.exit(1);
+        return flushAndExit(1);
       } else {
         peerId = peerIdR.value;
       }
@@ -613,8 +618,7 @@ app
 
     if (startupCheck.anyFailedChecks()) {
       logger.fatal({ reason: "Startup checks failed" }, "shutting down hub");
-      logger.flush();
-      process.exit(1);
+      return flushAndExit(1);
     }
 
     // Opt-out Diagnostics Reporting
@@ -649,10 +653,9 @@ app
       if (!startupCheck.anyFailedChecks()) {
         logger.fatal(hubResult.error);
         logger.fatal({ reason: "Hub Creation failed" }, "shutting down hub");
-
-        logger.flush();
       }
-      process.exit(1);
+
+      return flushAndExit(1);
     }
 
     if (statsDServer && !disableConsoleStatus) {
@@ -702,8 +705,8 @@ app
       try {
         await hub.teardown(HubShutdownReason.EXCEPTION);
       } finally {
-        logger.flush();
-        process.exit(1);
+        // Using return here would be unsafe
+        flushAndExit(1);
       }
     }
 
@@ -721,18 +724,18 @@ const s3SnapshotURL = new Command("snapshot-url")
     const network = farcasterNetworkFromJSON(options.network ?? FarcasterNetwork.MAINNET);
     if (network !== FarcasterNetwork.MAINNET) {
       console.error("Only mainnet snapshots are supported at this time");
-      exit(1);
+      return flushAndExit(1);
     }
 
     const response = await snapshotURLAndMetadata(network, 0, options.s3SnapshotBucket);
     if (response.isErr()) {
       console.error("error fetching snapshot data", response.error);
-      exit(1);
+      return flushAndExit(1);
     }
     const [url, metadata] = response.value;
     console.log(`${JSON.stringify(metadata, null, 2)}`);
     console.log(`Download chunks under directory at: ${url}`);
-    exit(0);
+    return flushAndExit(1);
   });
 
 app.addCommand(s3SnapshotURL);
@@ -779,7 +782,7 @@ const createIdCommand = new Command("create")
       await writePeerId(peerId, resolve(path));
     }
 
-    exit(0);
+    return flushAndExit(0);
   });
 
 const verifyIdCommand = new Command("verify")
@@ -788,7 +791,7 @@ const verifyIdCommand = new Command("verify")
   .action(async (options) => {
     const peerId = await readPeerId(options.id);
     logger.info(`Successfully Read peerId: ${peerId.toString()} from ${options.id}`);
-    exit(0);
+    return flushAndExit(0);
   });
 
 app
@@ -823,7 +826,7 @@ app
         "The 'status' command has been deprecated\n" +
         "Please use Grafana monitoring. See https://www.thehubble.xyz/intro/monitoring.html\n",
     );
-    exit(0);
+    return flushAndExit(0);
   });
 
 /*//////////////////////////////////////////////////////////////
@@ -849,7 +852,7 @@ const storageProfileCommand = new Command("storage")
     }
 
     await rocksDB.close();
-    exit(0);
+    return flushAndExit(0);
   });
 
 const rpcProfileCommand = new Command("rpc")
@@ -916,7 +919,7 @@ app
     }
 
     logger.info({ rocksDBName }, "Database cleared.");
-    exit(0);
+    return flushAndExit(0);
   });
 
 /*//////////////////////////////////////////////////////////////