+++ packages/core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @farcaster/core
 
+## 0.14.15
+
+### Patch Changes
+
+- c424e111: feat: Add support for long casts
+
+## 0.14.14
+
+### Patch Changes
+
+- de52fbce: feat: extend verifications signature max length
+
 ## 0.14.13
 
 ### Patch Changes