+++ packages/shuttle/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @farcaster/hub-shuttle
 
+## 0.3.3
+
+### Patch Changes
+
+- fix: Fix shuttle hub-nodejs version dep to support long casts
+
+## 0.3.2
+
+### Patch Changes
+
+- feat(shuttle): store cast type in message body
+
+## 0.3.1
+
+### Patch Changes
+
+- 87796697: Export multiple signatures for farcasterTimeToDate
+  - @farcaster/hub-nodejs@0.11.15
+
 ## 0.3.0
 
 ### Minor Changes